Allow reusing an external AST
If the user needs to perform additional processing on the AST using some tool other than falafel, it is useful to just give falafel an already existing AST.

This commit makes falafel look for an `ast` property in `opts` before invoking the `parser`. Note that providing the source code is still necessary for falafel to work properly.